feat(redhat-system): removing epel repository
Extra Packages for Enterprise Linux (or EPEL)
is a Fedora Special Interest Group that creates,
maintains, and manages a high quality set of
additional packages for Enterprise Linux.

This repos was used to install ansible via
"yum install" command, this repos is not include
by default in redhat distribution so it should not
be present in a "base" docker image.

This commit remove epel repository and install
ansible with python-pip.