feat(redhat-system): upgrade python-pip
The installiation of the latest version
of Ansible can cause an error on Arm
architecture during the compilation
of the Cryptographie library dependency.
To avoid the compilation of Python
cryptographie, the latest version of
python-pip must be installed.

This commit upgrade python-pip on redhat family
images to avoid compilation of Cryptography lib
during Ansible install.