fix(archlinux): reduce size of docker steps
The Archlinux image could not be build with
docker buildx because one of the dockerfile
step wass to too heavy, docker buildx returned
an error "no enough free disk space". Doesn't
know why its returned this error.

This commit fix that by reduced the size of the
dockerfile steps by moving the ansible installation
in a dedicated step.