feat(os): support rockylinux image
Rocky Linux is a community enterprise operating system designed to be
100% bug-for-bug compatible with Red Hat Linux.

The CentOS project recently announced a shift in strategy for CentOS.
Whereas previously CentOS existed as a downstream build of its upstream
vendor (it receives patches and updates after the upstream vendor does),
it will be shifting to an upstream build (testing patches and updates
before inclusion in the upstream vendor). Additionally, support for
CentOS Linux 8 has been cut short, from May 31, 2029 to December 31, 2021.

This commit adds rockylinux image with Systemd and Ansible installed.